+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Changelog
 
 ## UNRELEASED
+
+## v0.12 (2025-04-04)
 - bumped required demosplan-addon version to v0.51
 - fetch orga by 401 included name and set its planners as authorized
 - update the standard to 1.3 (new xsd files), mapping of every xsd namespace to
- on php namespace
+  on php namespace
 - generate new schema yml files and php classes from xsd files
 - Update the current addon code to be functional with the new 1.3 standard.
 - Update XBeteiligungService